+++ lib/dradis/plugins/veracode/field_processor.rb
@@ -2,17 +2,22 @@ module Dradis
   module Plugins
     module Veracode
       class FieldProcessor < Dradis::Plugins::Upload::FieldProcessor
-        def post_initialize(args={})
+        def post_initialize(args = {})
+          @record =
+            if (data.is_a?(::Veracode::Flaw) || data.is_a?(::Veracode::Evidence))
+              data
 
-          # Dealing with XML from Plugin Manager
-          if (data.name == 'cwe')
-            @flaw = ::Veracode::Flaw.new(data.at_xpath('./staticflaws/flaw'))
-          else
-            @flaw = ::Veracode::Flaw.new(data)
-          end
+            # Note: The evidence and flaw samples are the same but they need to
+            # be differentiated in the plugins manager preview. In that case,
+            # we're adding a "dradis_type" attribute in the evidence.sample file
+            elsif (data['dradis_type'] == 'evidence')
+              ::Veracode::Evidence.new(data.at_xpath('./staticflaws/flaw'))
+            else
+              ::Veracode::Flaw.new(data.at_xpath('./staticflaws/flaw'))
+            end
         end
 
-        def value(args={})
+        def value(args = {})
           field = args[:field]
 
           # fields in the template are of the form <template>.<name>, where
@@ -20,7 +25,7 @@ module Dradis
           # meaningless).
           _, name = field.split('.')
 
-          @flaw.try(name) || 'n/a'
+          @record.try(name) || 'n/a'
         end
       end
     end