+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 v4.12.0 (XXXX 2024)
+  - Create evidence for every instance of <flaw>
+  - Use cweid as the issue identifier
   - Update Dradis links in README
 
 v4.11.0 (January 2024)